op-node: move MeteredL1Fetcher
